--- docs/src/main/java/io/helidon/docs/includes/guides/lc4j/memory/PirateMemoryProvider.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2025 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.docs.includes.guides.lc4j.memory;
-// tag::snippet_1[]
-import java.util.function.Supplier;
-
-import io.helidon.service.registry.Service;
-
-import dev.langchain4j.memory.chat.ChatMemoryProvider;
-import dev.langchain4j.memory.chat.MessageWindowChatMemory;
-import dev.langchain4j.store.memory.chat.InMemoryChatMemoryStore;
-
-@Service.Singleton
-@Service.Named(PirateMemoryProvider.NAME)
-public class PirateMemoryProvider implements Supplier<ChatMemoryProvider> {
-
-    static final String NAME = "pirate-memory";
-
-    @Override
-    public ChatMemoryProvider get() {
-        return memoryId -> MessageWindowChatMemory.builder()
-                .maxMessages(10)
-                .id(memoryId)
-                .chatMemoryStore(new InMemoryChatMemoryStore()).build();
-    }
-}
-// end::snippet_1[]
