--- docs/src/main/java/io/helidon/docs/se/guides/lc4j/memory/Resource.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2025 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.docs.se.guides.lc4j.memory;
-
-import io.helidon.docs.includes.guides.lc4j.memory.PirateService;
-// tag::imports[]
-import io.helidon.http.HeaderNames;
-import io.helidon.service.registry.Services;
-import io.helidon.webserver.http.HttpRouting;
-// end::imports[]
-
-public class Resource {
-
-    // tag::method[]
-    static void routing(HttpRouting.Builder routing) {
-        routing.post("/chat", (req, res) -> {
-            var prompt = req.content().as(String.class);
-            var conversationId = req.headers()
-                    .get(HeaderNames.create("conversation-id"))
-                    .getString();
-            var response = Services.get(PirateService.class)
-                    .chat(conversationId, "Frank", prompt);
-
-            res.send(response);
-        });
-    }
-    // end::method[]
-}
